+++ project/app/src/oss/java/org/owntracks/android/location/NoopActivityRecognitionClient.kt
@@ -1,0 +1,11 @@
+package org.owntracks.android.location
+
+import timber.log.Timber
+
+class NoopActivityRecognitionClient : ActivityRecognitionClient {
+  override fun requestActivityUpdates() {
+    Timber.d("Activity recognition is unavailable without Google Play Services")
+  }
+
+  override fun removeActivityUpdates() {}
+}